+++ src/app/modules/cart/cart-item/cart-item.component.ts
@@ -9,7 +9,7 @@ import { IProduct } from '../../shared/interfaces';
 })
 export class CartItemComponent implements OnInit {
   @Input() product: IProduct;
-  @Output() changedQty: EventEmitter<IProduct> = new EventEmitter<IProduct>();
+  @Output() changedQty: EventEmitter<{product: IProduct, qty: number}> = new EventEmitter<{product: IProduct, qty: number}>();
   @Output() removed: EventEmitter<IProduct> = new EventEmitter<IProduct>();
 
   @HostBinding('class') attrClass;
@@ -30,8 +30,7 @@ export class CartItemComponent implements OnInit {
   }
 
   onChangeQty(qty: number): void {
-    this.product.quantity = +qty;
-    this.changedQty.emit(this.product);
+    this.changedQty.emit({product: this.product, qty: +qty});
   }
 
   remove() {