+++ src/app/modules/core/services/generete-n-sequence.factory.ts
@@ -1,0 +1,4 @@
+export const generateRandomSequence = (n: number): string => {
+  return new Array(n).join().replace(/(.|$)/g, function() { return ((Math.random() * 36) | 0)
+    .toString(36)[Math.random() < .5 ? 'toString' : 'toUpperCase'](); });
+};