+++ src/app/modules/products/products.module.ts
@@ -5,17 +5,20 @@ import { SharedModule } from '../shared/shared.module';
 import { ProductListComponent } from './product-list/product-list.component';
 import { ProductComponent } from './product-list/product/product.component';
 import { ProductService } from './products.service';
+import { TestCoreServicesComponent } from './test-core-services/test-core-services.component';
 
 @NgModule({
   declarations: [
     ProductListComponent,
-    ProductComponent
+    ProductComponent,
+    TestCoreServicesComponent
   ],
   imports: [
     SharedModule
   ],
   exports: [
-    ProductListComponent
+    ProductListComponent,
+    TestCoreServicesComponent
   ],
   providers: [ProductService]
 })